fix(UploadPicker): Show menu if folders are enabled
If folders are enabled but no "new"-menu entries, we need to also show the menu,
to allow users to choose between folder and file upload.
