+++ cypress/components/UploadPicker/UploadPicker.cy.ts
@@ -113,93 +113,6 @@ describe('UploadPicker valid uploads', () => {
 	})
 })
 
-describe('NewFileMenu handling', () => {
-	const propsData = {
-		destination: new Folder({
-			id: 56,
-			owner: 'user',
-			source: generateRemoteUrl('dav/files/user/Folder'),
-			permissions: Permission.ALL,
-			root: '/files/user',
-		}),
-	}
-	const entry = {
-		id: 'empty-file',
-		displayName: 'Create empty file',
-		templateName: 'New file',
-		iconClass: 'icon-file',
-		enabled: (folder: Folder) => (folder.permissions & Permission.CREATE) !== 0,
-		handler() {},
-	} as Entry
-
-	before(() => {
-		cy.spy(entry, 'handler')
-		addNewFileMenuEntry(entry)
-	})
-
-	afterEach(() => resetDocument())
-
-	it('Open the New File Menu', () => {
-		// Mount picker
-		cy.mount(UploadPicker, { propsData })
-
-		// Check and init aliases
-		cy.get('[data-cy-upload-picker] [data-cy-upload-picker-input]').as('input').should('exist')
-		cy.get('[data-cy-upload-picker] .upload-picker__progress').as('progress').should('exist')
-		cy.get('[data-cy-upload-picker] .action-item__menutoggle')
-			.as('menuButton')
-			.should('exist')
-
-		cy.get('@menuButton').click()
-		cy.get('[data-cy-upload-picker-menu-entry="upload-file"]').should('have.length', 1)
-		cy.get('[data-cy-upload-picker-menu-entry="empty-file"]').should('have.length', 1)
-
-		cy.get('[data-cy-upload-picker-menu-entry="empty-file"]')
-			.click()
-			.then(() => {
-				expect(entry.handler).to.be.called
-			})
-	})
-
-	it('Changes the context', () => {
-		// Mount picker
-		cy.mount(UploadPicker, { propsData })
-
-		// Check and init aliases
-		cy.get('[data-cy-upload-picker] [data-cy-upload-picker-input]').as('input').should('exist')
-		cy.get('[data-cy-upload-picker] .upload-picker__progress').as('progress').should('exist')
-		cy.get('[data-cy-upload-picker] .action-item__menutoggle')
-			.as('menuButton')
-			.should('exist')
-
-		cy.get('@menuButton').click()
-		cy.get('[data-cy-upload-picker-menu-entry="upload-file"]').should('have.length', 1)
-		cy.get('[data-cy-upload-picker-menu-entry="empty-file"]').should('have.length', 1)
-
-		// Close menu
-		cy.get('body').click()
-		cy.get('[data-cy-upload-picker-menu-entry="upload-file"]').should('not.be.visible')
-		cy.get('[data-cy-upload-picker-menu-entry="empty-file"]').should('not.be.visible')
-
-		cy.get('@component').then((component) => {
-			component.setDestination(new Folder({
-				id: 56,
-				owner: 'user',
-				source: generateRemoteUrl('dav/files/user/Folder'),
-				permissions: Permission.NONE,
-				root: '/files/user',
-			}))
-		})
-
-		// Menu should not be visible anymore
-		cy.get('[data-cy-upload-picker] .action-item__menutoggle')
-			.as('menuButton')
-			.should('not.exist')
-		cy.get('[data-cy-upload-picker-menu-entry="upload-file"]').should('have.length', 1)
-		cy.get('[data-cy-upload-picker-menu-entry="empty-file"]').should('not.exist')
-	})
-})
-
 describe('UploadPicker valid uploads', () => {
 	beforeEach(() => {
 		// Make sure we reset the destination
